+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.5.0 PropTypes as external package
+
+* react 15.5.x deprecates `PropTypes`: user NPM's `prop-types`.
+
 # 2.4.0 cordova friendly
 
 * Avoid analyzing cordova's hot code push URL: `/__cordova/`.